requirements.txt
aiodns==2.0.0
aiofiles==0.6.0
aiohttp==3.7.3
altgraph==0.17
async-spotify==0.4.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
dataclasses-json==0.5.2
discord-ext-ipc==2.0.0
discord-py-slash-command==1.0.8.4
discord.py==1.6.0
future==0.18.2
h11==0.12.0
h2==4.0.0
helpdev==0.7.1
hpack==4.0.0
httpcore==0.12.3
httpx==0.17.0
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
insprio==0.0.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
lavalink==3.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
mss==6.1.0
multidict==4.7.6
mutagen==1.45.1
mypy-extensions==0.4.3
numpy==1.20.0
oauthlib==3.1.0
opuslib==3.0.1
pandas==1.2.1
peewee==3.13.3
pefile==2019.4.18
priority==1.3.0
pycares==3.1.1
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.11
Pymem==1.4
PyNaCl==1.3.0
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.15.1.1.7.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-youtube==0.7.0
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
QDarkStyle==2.8.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
QtPy==1.9.0
Quart==0.14.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
sentry-sdk==0.19.4
six==1.15.0
sniffio==1.2.0
soundcloud-lib==0.5.4
soupsieve==2.2
spotify-dl==7.5.0
spotipy==2.16.1
stringcase==1.2.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
wavelink==0.9.9
Werkzeug==1.0.1
wsproto==1.0.0
yarl==1.5.1
youtube-data-api==0.0.20
youtube-dl==2021.2.4.1
youtube-search-python==1.4.1
